+++ apps/soritor/src/hooks/mutations/useMutationSignup.ts
@@ -1,20 +1,11 @@
-import { useMutation, useQueryClient } from "@tanstack/react-query";
+import { useMutation } from "@tanstack/react-query";
 
 import { signup } from "@/api/auth";
 
 import { FormSchemaType } from "@/hooks/useStackForm";
 
-import { AuthResponse } from "@/types/authType";
-
-import { setAccessToken } from "@/utils/handleToken";
-import { setRefreshToken } from "@/utils/handleCookie";
-
-import { user } from "../queries/user";
-
 
 export const useMutationSignup = () => {
-  const queryClient = useQueryClient();
-
   const mutation = useMutation({
     mutationFn: ({ userType, userName, userPhone }: Partial<FormSchemaType>) =>
       signup({
@@ -22,13 +13,6 @@ export const useMutationSignup = () => {
         userName,
         userPhone,
       }),
-    onSettled: () => {
-      queryClient.invalidateQueries({ queryKey: user.info().queryKey });
-    },
-    onSuccess: ({ accessToken, refreshToken }: AuthResponse) => {
-      setAccessToken(accessToken);
-      setRefreshToken("refreshToken", refreshToken);
-    },
   });
 
   return mutation;
