+++ apps/soritor/src/hooks/mutations/useMutationUpdateInfo.ts
@@ -1,22 +1,13 @@
 import { useMutation } from "@tanstack/react-query";
-import { useQueryClient } from "@tanstack/react-query";
 
 import { updateUserInfo } from "@/api/user";
 
-import { UserInfoResponse, UserInfoUpdateRequest } from "@/types/userType";
-
-import { user } from "../queries/user";
+import { UserInfoUpdateRequest } from "@/types/userType";
 
 
 export const useMutationUpdateInfo = () => {
-  const queryClient = useQueryClient();
-
   const mutation = useMutation({
     mutationFn: (userInfo: UserInfoUpdateRequest) => updateUserInfo(userInfo),
-    onSuccess: (data: UserInfoResponse) => {
-      queryClient.invalidateQueries({ queryKey: user.info().queryKey });
-      return data;
-    },
   });
 
   return mutation;
