+++ apps/soritor/src/hooks/useTicketStackForm.ts
@@ -1,7 +1,9 @@
 import { z } from "zod";
 import { UseFormReturn, useForm } from "react-hook-form";
+import { useQueryClient } from "@tanstack/react-query";
 import { zodResolver } from "@hookform/resolvers/zod";
 
+import { user } from "./queries/user";
 import { useMutationBuyTicket } from "./mutations/useMutationBuyTicket";
 
 export type FormSchemaType = z.infer<typeof FormSchema>;
@@ -13,6 +15,7 @@ export const FormSchema = z.object({
 });
 
 export const useTicketStackForm = () => {
+  const queryClient = useQueryClient();
   const { mutateAsync, isPending } = useMutationBuyTicket();
 
   const form = useForm<FormSchemaType>({
@@ -27,7 +30,15 @@ export const useTicketStackForm = () => {
   const onSubmit = async (data: FormSchemaType) => {
     const { universityId, reservationId } = data;
 
-    const response = await mutateAsync({ universityId, reservationId });
+    const response = await mutateAsync(
+      { universityId, reservationId },
+      {
+        onSuccess: data => {
+          queryClient.invalidateQueries({ queryKey: user.ticket().queryKey });
+          return data;
+        },
+      },
+    );
     return response.ticket;
   };
